### What I did

Before:

```
In [1]: %time import ape_arbitrum
CPU times: user 1.81 s, sys: 183 ms, total: 1.99 s
Wall time: 2.02 s
```

After:

```
CPU times: user 7.89 ms, sys: 3.25 ms, total: 11.1 ms
Wall time: 13.1 ms
```

### How I did it

### How to verify it

### Checklist

- [ ] Passes all linting checks (pre-commit and CI jobs)
- [ ] New test cases have been added and are passing
- [ ] Documentation has been updated
- [ ] PR title follows [Conventional Commit](https://www.conventionalcommits.org/en/v1.0.0/) standard (will be automatically included in the changelog)